Allow deleting all only on test profile

+++ src/main/java/dvarubla/tracker/api/AllController.java
@@ -1,5 +1,6 @@
 package dvarubla.tracker.api;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,12 +12,17 @@ import javax.transaction.Transactional;
 @RestController
 @RequestMapping(value = "/api/all")
 public class AllController {
+    @Value("${spring.profiles.active}")
+    private String activeProfile;
+
     @PersistenceContext
     private EntityManager _entityManager;
 
     @DeleteMapping
     @Transactional
     public void delete(){
-        _entityManager.createStoredProcedureQuery("delete_all").execute();
+        if(activeProfile.equals("test")) {
+            _entityManager.createStoredProcedureQuery("delete_all").execute();
+        }
     }
 }
